bootqca: clear the MBN header information only if authentication is successful
During secure boot, after the kernel image authentication, before knowing
the authentication status, we are clearing out the MBN information in the DDR.

When we try to boot the unsigned kernel image on secure board, since
there is no MBN header available, we try to access the invalid memory
region (or) size which leads to data abort (or) system reset without any failure message.

So clear the MBN information only if the image authentication is successful.
If the image authentication is failed, corresponding failure message will be
printed and system will go for reset.

Change-Id: I2cf8e098cdb7e0fc11d0b563d4a0d6a5f0d44ee8
Signed-off-by: Kathiravan T <quic_kathirav@quicinc.com>