board: arm: ipq5018: Add atf crypto fixup support
This changes fixup to '0' in "qcom,controlled-remotely"
property and '2' in "qti,config-pipe-trust-reg" property
in the crypto bam node if ATF is enabled. This will enable
the kernel to do complete bam pipe initialization.

And also this change update the ATF fixup api to
generic name so which will be common for all other
fixup releated to ATF.

Change-Id: I10b2e712883f419433220547a0d7bc6bdf4e125d
Signed-off-by: Vandhiadevan Karunamoorthy <quic_vkarunam@quicinc.com>