pack_v2: updated wifi_fw logic for pack script
This patch updates the following things.
1. Simplified the wififw pack script logic
2. Added the multi fw version support with simple changes

i. To add RDP specific fw img, mention the corresponding fw
name under each RDP entry in config.xml like below.

<wififw_name>file-name</wififw_name>

ii. When we migrate to new fw version for a specfic chipset,
add the following entry in the config.xml.

<data type="VERSION_PARAMETER">
	<version_check>chip-set-number</version_check>
	<default_version>base-fw-version</default_version>
</data>

Note: Also, make sure that all the fw names mentioned in the
config.xml are in the base verison.

iii. Once migration is done, remove the entries added in (ii).
and rename the new version fw names in the config.xml as
mentioned in the (i).

Change-Id: I296dab47e9ff321a1ed44130df8fc5cdd8a5bfa5
Signed-off-by: Ram Kumar D <quic_ramd@quicinc.com>